feat: généraliser l'export image au niveau des widgets cartographie
- Extrait la logique de capture dans un hook `useExportImage`
- Déplace `ModeExportContext` dans `_commons/Widget/` (les widgets ne dépendent plus de `ComparaisonTerritoires`)
- Crée `ExportableWidget`, wrapper générique ajoutant les boutons d'export à n'importe quel widget
- Refactorise `PanneauCarte` pour déléguer à `ExportableWidget`
- Wrappe `WidgetCartographieTA` standalone dans `PageChantiers` avec `ExportableWidget`
